finished task terminate logic Signed-off-by: Gavin520 <meiyan532484710@163.com>

+++ types/resource.go
@@ -55,8 +55,8 @@ func (m *Resource) GetDataStatus() apicommonpb.DataStatus { return m.data.DataSt
 func (m *Resource) GetState() apicommonpb.PowerState      { return m.data.State }
 func (m *Resource) GetTotalMem() uint64                   { return m.data.TotalMem }
 func (m *Resource) GetUsedMem() uint64 { return m.data.UsedMem }
-func (m *Resource) GetTotalProcessor() uint64 { return uint64(m.data.TotalProcessor) }
-func (m *Resource) GetUsedProcessor() uint64 { return uint64(m.data.UsedProcessor) }
+func (m *Resource) GetTotalProcessor() uint32 { return m.data.TotalProcessor }
+func (m *Resource) GetUsedProcessor() uint32 { return m.data.UsedProcessor }
 func (m *Resource) GetTotalBandWidth() uint64 { return m.data.TotalBandwidth }
 func (m *Resource) GetUsedBandWidth() uint64 { return m.data.UsedBandwidth }
 func (m *Resource) String() string {
